build: enable ko build [1/1]
PD#SWPL-72345

Problem:
enable ko build

Solution:
enable ko build

Verify:
on C2_AF400

+++ drivers/gki_tool/gki_module.c
@@ -216,7 +216,7 @@ void gki_module_init(void)
 			strcpy(cmdline, bootargs);
 	}
 
-	pr_debug("cmdline: %lx, %s\n", cmdline, cmdline);
+	pr_debug("cmdline: %lx, %s\n", (unsigned long)cmdline, cmdline);
 	cmdline_parse_args(cmdline);
 }
